Allow localThreshold to be configured
As its required when creating the MongoClient, it means the configuration has
to be passed to the DefaultMongoClientFactory.

MongoConnector now takes an options map so localThreshold can be passed to the DefaultMongoClientFactory.
The ReadConfig and WriteConfig have been refactored so they can optionally include the URI. Allowing for
easy creation of a Read or Write MongoConnector.

SPARK-29